feat(share_plus): export XFile
