+++ packages/share_plus/share_plus/example/integration_test/share_plus_test.dart
@@ -6,7 +6,6 @@
 
 import 'dart:io';
 
-import 'package:file_selector/file_selector.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:share_plus/share_plus.dart';
