+++ packages/share_plus/share_plus/example/lib/main.dart
@@ -6,10 +6,12 @@
 
 import 'dart:io';
 
-import 'package:file_selector/file_selector.dart';
+import 'package:file_selector/file_selector.dart'
+    hide XFile; // hides to test if share_plus exports XFile
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:image_picker/image_picker.dart';
+import 'package:image_picker/image_picker.dart'
+    hide XFile; // hides to test if share_plus exports XFile
 import 'package:share_plus/share_plus.dart';
 
 import 'image_previews.dart';
