+++ packages/share_plus/share_plus_platform_interface/lib/method_channel/method_channel_share.dart
@@ -9,7 +9,6 @@ import 'dart:io';
 // ignore: unnecessary_import
 import 'dart:ui';
 
-import 'package:cross_file/cross_file.dart';
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart' show visibleForTesting;
 import 'package:mime/mime.dart' show extensionFromMime, lookupMimeType;
