+++ packages/share_plus/share_plus_platform_interface/test/share_plus_platform_interface_test.dart
@@ -4,7 +4,6 @@
 
 import 'dart:io';
 
-import 'package:cross_file/cross_file.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart'
     show TestDefaultBinaryMessengerBinding, TestWidgetsFlutterBinding;
